(1) The channel tree view displays all NCs, even those that are not registered in the naming service.   This addresses an issue from JIRA:COMP-8997, because system NCs for alarm, logging etc are not registered there. Unfortunately they also don't carry a channel name other than the Id. - EventData#discoverChannels now merges NC information from (a) the naming service, (b) the notify service's regular API, (c), the MC TAO extension.   This is complicated because the overlap of information among these sources is not large, but care was taken to avoid unnecessary remote calls. - NCs not present in the naming service cannot be subscribed to, but they should eventually be equal to normal NCs in terms of subscriber monitoring etc. - ChannelData now carries the NC ID in addition to the NC name, to support merging NC information from different sources.
(2) code clean-up:
- Moved code from AbstractNotifyServiceElement to ChannelData, which was not applicable for the other subclass NotifyServiceData.
- Removed some redundancies between MCStatistics and its subclasses; INames interface no longer needed to mark those statistics that have child nodes. 
- Fixed compiler warnings.

git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@188551 523d945c-050c-4681-91ec-863ad3bb968a

+++ LGPL/CommonSoftware/acsGUIs/eventGUI/src/alma/acs/eventbrowser/model/EventData.java
@@ -68,7 +68,7 @@ public class EventData extends AbstractEventData {
 	 * @return Average number of events/s since the first one that was processed
 	 */
 	public static float getAverageRate() {
-		return ((float)getTotalEventsProcessed()*1000.f)/((float)(System.currentTimeMillis()-getTimeFirstEventProcessed()));
+		return (getTotalEventsProcessed()*1000.f)/(System.currentTimeMillis()-getTimeFirstEventProcessed());
 	}
 
 	public String getSourceObject() {
